--- Puzzel_src/api/sagas.js
@@ -1,6 +1,0 @@
-import {all} from 'redux-saga/effects';
-import {watchGetStatus} from '../controler/user.saga';
-
-export default function* rootSaga() {
-  yield all([watchGetStatus()]);
-}
