+++ babel.config.js
@@ -1,7 +1,3 @@
 module.exports = {
-  presets: [
-    'module:@react-native/babel-preset',
-    ['@babel/preset-env', {targets: {node: 'current'}}],
-     '@babel/preset-typescript'
-    ]
+  "presets": ["module:metro-react-native-babel-preset"]
 };
