--- eslint.config.js
@@ -1,11 +1,0 @@
-{
-    "overrides": [
-      {
-        "files": ["tests/**/*"],
-        "plugins": ["jest"],
-        "env": {
-          "jest/globals": true
-        }
-      }
-    ]
-  }
