+++ jest.config.js
@@ -1,6 +1,3 @@
 module.exports = {
   preset: 'react-native',
-  "transformIgnorePatterns": [
-    "/node_modules/(?!(@react-navigation/elements)/)"
-  ]
 };
