+++ package.json
@@ -14,7 +14,7 @@
   "dependencies": {
     "@react-native-async-storage/async-storage": "^1.21.0",
     "@react-native-firebase/app": "^19.1.2",
-    "@react-native-firebase/firestore": "^19.1.2",
+    "@react-native-firebase/messaging": "^19.2.1",
     "@react-navigation/native": "^6.1.10",
     "@react-navigation/native-stack": "^6.9.18",
     "@react-navigation/stack": "^6.3.21",
@@ -23,38 +23,33 @@
     "react": "18.2.0",
     "react-native": "0.73.4",
     "react-native-gesture-handler": "^2.15.0",
-    "react-native-gifted-chat": "^2.4.0",
     "react-native-mmkv": "^2.12.1",
     "react-native-safe-area-context": "^4.9.0",
     "react-native-screens": "^3.29.0",
     "react-native-sqlite-storage": "^6.0.1",
     "react-native-swipeable": "^0.6.0",
+    "react-query": "^3.39.3",
     "react-redux": "^9.1.0",
     "redux": "^5.0.1",
-    "redux-mock-store": "^1.5.4",
     "redux-persist": "^6.0.0",
     "redux-saga": "^1.3.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.24.4",
-    "@babel/preset-env": "^7.24.4",
-    "@babel/preset-typescript": "^7.24.1",
+    "@babel/core": "^7.20.0",
+    "@babel/preset-env": "^7.20.0",
     "@babel/runtime": "^7.20.0",
-    "@jest/globals": "^29.7.0",
     "@react-native/babel-preset": "0.73.21",
     "@react-native/eslint-config": "0.73.2",
     "@react-native/metro-config": "0.73.5",
     "@react-native/typescript-config": "0.73.1",
-    "@testing-library/react-native": "^12.4.5",
-    "@types/jest": "^29.5.12",
     "@types/react": "^18.2.6",
     "@types/react-test-renderer": "^18.0.0",
-    "babel-jest": "^29.7.0",
+    "babel-jest": "^29.6.3",
     "eslint": "^8.19.0",
     "jest": "^29.6.3",
+    "metro-react-native-babel-preset": "^0.77.0",
     "prettier": "2.8.8",
     "react-test-renderer": "18.2.0",
-    "ts-jest": "^29.1.2",
     "typescript": "5.0.4"
   },
   "engines": {
