+++ src/components/ColoredBetweenAsterisk.tsx
